+++ packages/juice/lib/src/bloc/src/use_case_builders/src/relay_use_case_builder.dart
@@ -104,6 +104,9 @@ class RelayUseCaseBuilder<
   Type get eventType => typeOfEvent;
 
   @override
+  EventConcurrency get concurrency => EventConcurrency.concurrent;
+
+  @override
   UseCaseGenerator get generator => useCaseGenerator;
 
   @override